Add Metadata DTO to Data package
