+++ app/src/main/java/com/sefford/artdrian/Artpplication.kt
@@ -1,6 +1,7 @@
 package com.sefford.artdrian
 
 import android.app.Application
+import com.sefford.artdrian.data.Endpoints
 import com.sefford.artdrian.di.ApplicationModule
 import com.sefford.artdrian.di.CoreModule
 import com.sefford.artdrian.di.DaggerApplicationComponent
