+++ app/src/main/java/com/sefford/artdrian/data/Endpoints.kt
@@ -1,4 +1,4 @@
-package com.sefford.artdrian
+package com.sefford.artdrian.data
 
 object Endpoints {
     const val ENDPOINT = "https://adrianmg-go-server.fly.dev/"
