+++ app/src/main/java/com/sefford/artdrian/wallpaperlist/ui/WallpaperListViewModel.kt
@@ -2,10 +2,10 @@ package com.sefford.artdrian.wallpaperlist.ui
 
 import androidx.lifecycle.ViewModel
 import arrow.core.Either
-import com.sefford.artdrian.datasources.WallpaperRepository
-import com.sefford.artdrian.datasources.WallpaperRepository.*
-import com.sefford.artdrian.datasources.WallpaperRepository.RepositoryError.NetworkingError
-import com.sefford.artdrian.datasources.WallpaperRepository.RepositoryError.NotFound
+import com.sefford.artdrian.data.datasources.WallpaperRepository
+import com.sefford.artdrian.data.datasources.WallpaperRepository.*
+import com.sefford.artdrian.data.datasources.WallpaperRepository.RepositoryError.NetworkingError
+import com.sefford.artdrian.data.datasources.WallpaperRepository.RepositoryError.NotFound
 import com.sefford.artdrian.model.Wallpaper
 import com.sefford.artdrian.usecases.GetWallpapers
 import kotlinx.coroutines.CoroutineDispatcher
