+++ app/src/test/java/com/sefford/artdrian/datasources/FakeWallpaperApi.kt
@@ -1,5 +1,6 @@
 package com.sefford.artdrian.datasources
 
+import com.sefford.artdrian.data.datasources.WallpaperApi
 import com.sefford.artdrian.model.Metadata
 
 class FakeWallpaperApi(private val response: () -> List<Metadata>) : WallpaperApi {
